V1: Occupancy-Conditioned Joint Trajectory-Geometry Diffusion (docs/联合扩散.md)
- 删除旧 z-残差/零和/顺序椭圆支路/guidance/AL 安全 loss 与三阶段训练
- 新模型 src/models/joint: SceneCNN 16x16 + phi/psi/type/time 编码 + w_t gate + 256 交错 token + 8x AdaLN JointBlock(相对bias self-attn+cross-attn+FFN) + 双 x0 head
- 数据: data/processed_scene_v1 (128 锚点含 E0, e6 表示, 去掉 valid 门控)
- train/sample/evaluate 重写: 共享 alpha-bar, x0-prediction + DDIM, p1=s/p128=g 端点硬条件
- 保留数据与几何工具 (scene_obstacles 迁移), docs/V1.md

+++ src/models/position_encoding.py
@@ -53,7 +53,7 @@ class Sinusoidal1DPositionEmbedding(nn.Module):
 
     def forward(self, pos):
         """pos [..., L] integers -> [..., L, d_model]."""
-        args = pos[..., :, None] * self.freqs                   # [..., L, half]
+        args = pos.float()[..., :, None] * self.freqs                    # [..., L, half]
         emb = torch.stack([torch.sin(args), torch.cos(args)], dim=-1)  # [..., L, half, 2]
         return emb.reshape(*emb.shape[:-2], -1)                 # [..., L, d_model]
 
@@ -74,8 +74,6 @@ class SinusoidalTimestepEmbedding(nn.Module):
         args = t[:, None] * self.freqs[None]
         emb = torch.cat([torch.cos(args), torch.sin(args)], dim=-1)
         return self.mlp(emb)
-
-
 class Sinusoidal2DRelativePositionEmbedding(nn.Module):
     """Relative 2D sinusoidal position embedding for small scene-space offsets.
